refactor: narrow ViewModel dependencies, fix package paths, and simplify UI
- Narrow SyncViewModel constructor dependencies from PlatformOperations to
  focused ClipboardOperations and FileOperations interfaces.
- Align AndroidDiscoveryService and DesktopDiscoveryService packages with
  their directory paths, moving factory implementations into dedicated
  NetworkDiscoveryService platform files.
- Replace internal kotlinx.coroutines tryResume calls in Android NSD
  resolution with standard resume API.
- Simplify connected screen UI by hiding self-device details and security
  banners once a peer connection is active.
- Remove redundant media picker actions, keeping files as the single
  unified option for file picking.
