+++ shared/src/androidMain/kotlin/com/liftley/sync360/core/di/Koin.android.kt
@@ -7,7 +7,7 @@ import com.liftley.sync360.core.platform.IncomingMessageNotifier
 import com.liftley.sync360.core.platform.PlatformOperations
 import com.liftley.sync360.features.sync.domain.model.DeviceProfile
 import com.liftley.sync360.features.sync.domain.model.createLocalDeviceProfile
-import com.liftley.sync360.features.sync.domain.network.AndroidDiscoveryService
+import com.liftley.sync360.features.sync.data.network.AndroidDiscoveryService
 import com.liftley.sync360.features.sync.domain.network.NetworkDiscoveryService
 import org.koin.core.module.Module
 import org.koin.dsl.module
