+++ shared/src/jvmMain/kotlin/com/liftley/sync360/features/sync/domain/network/NetworkDiscoveryService.jvm.kt
@@ -1,0 +1,7 @@
+package com.liftley.sync360.features.sync.domain.network
+
+import com.liftley.sync360.features.sync.data.network.DesktopDiscoveryService
+
+actual fun createNetworkDiscoveryService(context: Any?): NetworkDiscoveryService {
+    return DesktopDiscoveryService()
+}
